Default to live mode for empty settings.
May fix Broken Preview Mode? #29
It appears that it is possible for the config.livePreview to be undefined for new vaults. On my machine, the contents of app.json on a new vault is an empty object. Therefore, by default config.livePreview is undefined, and only is set if you alter the default settings. This change correctly sets the livePreview mode to true if it is undefined in the config.